+++ lambdas/issuecredential/src/test/java/uk/gov/di/ipv/cri/address/api/service/VerifiableCredentialServiceTest.java
@@ -15,9 +15,9 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.di.ipv.cri.address.api.service.fixtures.TestFixtures;
-import uk.gov.di.ipv.cri.address.library.domain.CanonicalAddress;
-import uk.gov.di.ipv.cri.address.library.service.ConfigurationService;
-import uk.gov.di.ipv.cri.address.library.util.SignedJWTFactory;
+import uk.gov.di.ipv.cri.common.library.domain.CanonicalAddress;
+import uk.gov.di.ipv.cri.common.library.service.ConfigurationService;
+import uk.gov.di.ipv.cri.common.library.util.SignedJWTFactory;
 
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
